+++ gems/aws-sdk-cloudformation/lib/aws-sdk-cloudformation/resource.rb
@@ -353,6 +353,21 @@ module Aws::CloudFormation
     #   })
     # @param [Hash] options ({})
     # @option options [String] :stack_name
+    #   <note markdown="1"> If you don't pass a parameter to `StackName`, the API returns a
+    #   response that describes all resources in the account. This requires
+    #   `ListStacks` and `DescribeStacks` permissions.
+    #
+    #    The IAM policy below can be added to IAM policies when you want to
+    #   limit resource-level permissions and avoid returning a response when
+    #   no parameter is sent in the request:
+    #
+    #    \\\{ "Version": "2012-10-17", "Statement": \[\\\{ "Effect":
+    #   "Deny", "Action": "cloudformation:DescribeStacks",
+    #   "NotResource": "arn:aws:cloudformation:*:*:stack/*/*" \\}\]
+    #   \\}
+    #
+    #    </note>
+    #
     #   The name or the unique stack ID that's associated with the stack,
     #   which aren't always interchangeable:
     #