+++ gems/aws-sdk-cloudformation/lib/aws-sdk-cloudformation/stack.rb
@@ -195,7 +195,7 @@ module Aws::CloudFormation
     end
 
     # Information about whether a stack's actual configuration differs, or
-    # has *drifted*, from it's expected configuration, as defined in the
+    # has *drifted*, from its expected configuration, as defined in the
     # stack template and any values specified as template parameters. For
     # more information, see [Detecting Unregulated Configuration Changes to
     # Stacks and Resources][1].